[refactor] AttachmentService 하위에 있던 setUserProfileImage 메서드 UserService 클래스 하위 메서드로 변경
